requirements.txt
Flask == 2.2.5
africastalking == 1.2.5
python-dotenv = 1.0.0
Flask-Cors == 4.0.0